+++ tests/IO/EwktWriterTest.php
@@ -21,8 +21,7 @@ class EwktWriterTest extends EwktAbstractTestCase
     #[DataProvider('providerPrettyPrint')]
     public function testPrettyPrint(bool $prettyPrint, string $ewkt) : void
     {
-        $writer = new EwktWriter();
-        $writer->setPrettyPrint($prettyPrint);
+        $writer = new EwktWriter(prettyPrint: $prettyPrint);
 
         $lineString = $this->createLineString([[1, 2, 3, 4], [5, 6, 7, 8]], CoordinateSystem::xyzm(4326));
 
@@ -46,8 +45,7 @@ class EwktWriterTest extends EwktAbstractTestCase
     #[DataProvider('providerPointWkt')]
     public function testWritePoint(string $wkt, array $coords, bool $is3D, bool $isMeasured) : void
     {
-        $writer = new EwktWriter();
-        $writer->setPrettyPrint(false);
+        $writer = new EwktWriter(prettyPrint: false);
 
         $cs = new CoordinateSystem($is3D, $isMeasured, 4326);
         $point = $this->createPoint($coords, $cs);
@@ -64,8 +62,7 @@ class EwktWriterTest extends EwktAbstractTestCase
     #[DataProvider('providerLineStringWkt')]
     public function testWriteLineString(string $wkt, array $coords, bool $is3D, bool $isMeasured) : void
     {
-        $writer = new EwktWriter();
-        $writer->setPrettyPrint(false);
+        $writer = new EwktWriter(prettyPrint: false);
 
         $cs = new CoordinateSystem($is3D, $isMeasured, 4326);
         $lineString = $this->createLineString($coords, $cs);
@@ -82,8 +79,7 @@ class EwktWriterTest extends EwktAbstractTestCase
     #[DataProvider('providerCircularStringWkt')]
     public function testWriteCircularString(string $wkt, array $coords, bool $is3D, bool $isMeasured) : void
     {
-        $writer = new EwktWriter();
-        $writer->setPrettyPrint(false);
+        $writer = new EwktWriter(prettyPrint: false);
 
         $cs = new CoordinateSystem($is3D, $isMeasured, 4326);
         $lineString = $this->createCircularString($coords, $cs);
@@ -100,8 +96,7 @@ class EwktWriterTest extends EwktAbstractTestCase
     #[DataProvider('providerCompoundCurveWkt')]
     public function testWriteCompoundCurve(string $wkt, array $coords, bool $is3D, bool $isMeasured) : void
     {
-        $writer = new EwktWriter();
-        $writer->setPrettyPrint(false);
+        $writer = new EwktWriter(prettyPrint: false);
 
         $cs = new CoordinateSystem($is3D, $isMeasured, 4326);
         $compoundCurve = $this->createCompoundCurve($coords, $cs);
@@ -118,8 +113,7 @@ class EwktWriterTest extends EwktAbstractTestCase
     #[DataProvider('providerPolygonWkt')]
     public function testWritePolygon(string $wkt, array $coords, bool $is3D, bool $isMeasured) : void
     {
-        $writer = new EwktWriter();
-        $writer->setPrettyPrint(false);
+        $writer = new EwktWriter(prettyPrint: false);
 
         $cs = new CoordinateSystem($is3D, $isMeasured, 4326);
         $polygon = $this->createPolygon($coords, $cs);
@@ -136,8 +130,7 @@ class EwktWriterTest extends EwktAbstractTestCase
     #[DataProvider('providerTriangleWkt')]
     public function testWriteTriangle(string $wkt, array $coords, bool $is3D, bool $isMeasured) : void
     {
-        $writer = new EwktWriter();
-        $writer->setPrettyPrint(false);
+        $writer = new EwktWriter(prettyPrint: false);
 
         $cs = new CoordinateSystem($is3D, $isMeasured, 4326);
         $triangle = $this->createTriangle($coords, $cs);
@@ -154,8 +147,7 @@ class EwktWriterTest extends EwktAbstractTestCase
     #[DataProvider('providerCurvePolygonWkt')]
     public function testWriteCurvePolygon(string $wkt, array $coords, bool $is3D, bool $isMeasured) : void
     {
-        $writer = new EwktWriter();
-        $writer->setPrettyPrint(false);
+        $writer = new EwktWriter(prettyPrint: false);
 
         $cs = new CoordinateSystem($is3D, $isMeasured, 4326);
         $polygon = $this->createCurvePolygon($coords, $cs);
@@ -172,8 +164,7 @@ class EwktWriterTest extends EwktAbstractTestCase
     #[DataProvider('providerPolyhedralSurfaceWkt')]
     public function testWritePolyhedralSurface(string $wkt, array $coords, bool $is3D, bool $isMeasured) : void
     {
-        $writer = new EwktWriter();
-        $writer->setPrettyPrint(false);
+        $writer = new EwktWriter(prettyPrint: false);
 
         $cs = new CoordinateSystem($is3D, $isMeasured, 4326);
         $polyhedralSurface = $this->createPolyhedralSurface($coords, $cs);
@@ -190,8 +181,7 @@ class EwktWriterTest extends EwktAbstractTestCase
     #[DataProvider('providerTinWkt')]
     public function testWriteTin(string $wkt, array $coords, bool $is3D, bool $isMeasured) : void
     {
-        $writer = new EwktWriter();
-        $writer->setPrettyPrint(false);
+        $writer = new EwktWriter(prettyPrint: false);
 
         $cs = new CoordinateSystem($is3D, $isMeasured, 4326);
         $tin = $this->createTin($coords, $cs);
@@ -208,8 +198,7 @@ class EwktWriterTest extends EwktAbstractTestCase
     #[DataProvider('providerMultiPointWkt')]
     public function testWriteMultiPoint(string $wkt, array $coords, bool $is3D, bool $isMeasured) : void
     {
-        $writer = new EwktWriter();
-        $writer->setPrettyPrint(false);
+        $writer = new EwktWriter(prettyPrint: false);
 
         $cs = new CoordinateSystem($is3D, $isMeasured, 4326);
         $multiPoint = $this->createMultiPoint($coords, $cs);
@@ -226,8 +215,7 @@ class EwktWriterTest extends EwktAbstractTestCase
     #[DataProvider('providerMultiLineStringWkt')]
     public function testWriteMultiLineString(string $wkt, array $coords, bool $is3D, bool $isMeasured) : void
     {
-        $writer = new EwktWriter();
-        $writer->setPrettyPrint(false);
+        $writer = new EwktWriter(prettyPrint: false);
 
         $cs = new CoordinateSystem($is3D, $isMeasured, 4326);
         $multiLineString = $this->createMultiLineString($coords, $cs);
@@ -244,8 +232,7 @@ class EwktWriterTest extends EwktAbstractTestCase
     #[DataProvider('providerMultiPolygonWkt')]
     public function testWriteMultiPolygon(string $wkt, array $coords, bool $is3D, bool $isMeasured) : void
     {
-        $writer = new EwktWriter();
-        $writer->setPrettyPrint(false);
+        $writer = new EwktWriter(prettyPrint: false);
 
         $cs = new CoordinateSystem($is3D, $isMeasured, 4326);
         $multiPolygon = $this->createMultiPolygon($coords, $cs);
@@ -262,8 +249,7 @@ class EwktWriterTest extends EwktAbstractTestCase
     #[DataProvider('providerGeometryCollectionWkt')]
     public function testWriteGeometryCollection(string $wkt, array $coords, bool $is3D, bool $isMeasured) : void
     {
-        $writer = new EwktWriter();
-        $writer->setPrettyPrint(false);
+        $writer = new EwktWriter(prettyPrint: false);
 
         $cs = new CoordinateSystem($is3D, $isMeasured, 4326);